+++ jgclark.WindowSets/CHANGELOG.md
@@ -1,6 +1,9 @@
 # What's changed in 🖥️ Window Sets?
 _Please also see the Plugin [README](https://github.com/NotePlan/plugins/blob/main/jgclark.WindowSets/README.md)._
 
+## [0.3.0] - 2023-09-28 (unreleased; following change of design)
+Window Set definitions now live in a hidden datafile, _but can be automatically sync'd to/from a user's note to see what's going on_. See README for details.
+
 ## [0.2.0] - 2023-09-10 (unfinished; unreleased -- decided to change design)
 - Window Set definitions now live in notes. See README for details. Can now have any number of defined window sets.
 - **/Save window set** command. Now includes window size and position for floating windows. (Incomplete: doesn't yet work for updating existing Window Set definition.)